Remove achievement_name and attach achievement to task
Title now has the drop down with achievement names as suggestions
On using a title that matches an achievement, an achievement is
generated

By attaching the achievement to the task we should be able to remove
again when we change task to ongoing.

This is progressive enhancement.

The ultimate solution should be to

Merge Achievement model back into action plan task
Radio button to select achievement (no other option)
Update reports accordingly