Ugh ridiculous fix to avoid CV being downloaded on pressing enter
Seem the filterrific jquery searches for nearest button on pressing
enter to submit the form. As the csv download is the only button
this submit is triggered rather than reloading the page. I tried
suppressing the return key press from the search input but no
luck yet

This should be the correct way of fixing..